Remove TypeHelper. Move PlatformHelper to library module.
