--- packages/library/src/androidMain/kotlin/io/realm/internal/TypeFactory.kt
@@ -1,7 +1,0 @@
-package io.realm.internal
-
-import io.realm.log.RealmLogger
-
-actual object TypeFactory {
-    actual fun createDefaultSystemLogger(tag: String): RealmLogger = LogCatLogger(tag)
-}
