--- packages/library/src/commonMain/kotlin/io/realm/internal/TypeFactory.kt
@@ -1,7 +1,0 @@
-package io.realm.internal
-
-import io.realm.log.RealmLogger
-
-expect object TypeFactory {
-    fun createDefaultSystemLogger(tag: String): RealmLogger
-}
