+++ packages/library/src/jvmMain/kotlin/io/realm/internal/PlatformHelper.kt
@@ -1,0 +1,15 @@
+package io.realm.internal
+
+import io.realm.log.RealmLogger
+
+actual object PlatformHelper {
+    @Suppress("FunctionOnlyReturningConstant")
+    actual fun appFilesDirectory(): String {
+        // FIXME What is the standard default location for non-Android JVM builds.
+        //  https://github.com/realm/realm-kotlin/issues/75
+        return "."
+    }
+
+    actual fun createDefaultSystemLogger(tag: String): RealmLogger = StdOutLogger(tag)
+
+}
