+++ packages/library/src/macosMain/kotlin/io/realm/internal/PlatformHelper.kt
@@ -1,0 +1,12 @@
+package io.realm.internal
+
+import io.realm.log.RealmLogger
+
+actual object PlatformHelper {
+    @Suppress("FunctionOnlyReturningConstant")
+    actual fun appFilesDirectory(): String {
+        return "."
+    }
+
+    actual fun createDefaultSystemLogger(tag: String): RealmLogger = NSLogLogger(tag)
+}
